% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbmc2k.R
\name{LoadPBMC2k}
\alias{LoadPBMC2k}
\title{Load PBMC 2k example data}
\usage{
LoadPBMC2k(dev_path = NULL)
}
\arguments{
\item{dev_path}{Optional fallback path to pbmc2k.csv when running locally}
}
\value{
A Seurat object
}
\description{
Load Seurat object from pbmc2k.csv (used in HybridQC examples)
}
